+++ src/main/java/com/miaoyu/naigos/user/mapper/EditArchiveMapper.java
@@ -1,5 +1,5 @@
 package com.miaoyu.naigos.user.mapper;
 
 public interface EditArchiveMapper {
-    boolean editArchiveByUuid (String uuid, String nickname, String email, Integer qqId, String city);
+    boolean editArchiveByUuid (String uuid, String nickname, String email, Long qqId, String city);
 }
