+++ src/main/java/com/miaoyu/naigos/user/service/GetUserArchiveService.java
@@ -36,7 +36,7 @@ public class GetUserArchiveService {
         return switch (loginType) {
             case 1 -> getUserArchiveMapper.getUserArchiveByEmail(uid);
             case 2 -> getUserArchiveMapper.getUserArchiveByUuid(uid);
-            case 3 -> getUserArchiveMapper.getUserArchiveByQqId(Integer.valueOf(uid));
+            case 3 -> getUserArchiveMapper.getUserArchiveByQqId(Long.valueOf(uid));
             default -> null;
         };
     }
